+++ DummyOne/Goals/GoalKillSpecialGuard.cs
@@ -8,6 +8,7 @@ using OpenGlGameCommon.Classes;
 using SneakingCommon.Interfaces.Model;
 using SneakingCommon.System_Classes;
 using OpenGlGameCommon.Interfaces.Model;
+using SneakingCommon.Enums;
 
 namespace DummyOne.Goals
 {
@@ -20,7 +21,7 @@ namespace DummyOne.Goals
         }
         public bool goalReached(ArgOwner argOwner)
         {
-            List<IDrawableGuard> guards = (List < IDrawableGuard > )argOwner.getArg(ArgNames.guards);
+            List<IDrawableGuard> guards = (List < IDrawableGuard > )argOwner.getArg((int)ArgNames.guards);
             foreach (IDrawableGuard g in guards)
             {
                 if (g.getName() == guardName && g.MyCharacter.getValue("Is Dead") == 1)
